+++ CHANGELOG.md
@@ -46,7 +46,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- Fixed flaky integration test. ([#436](https://github.com/jina-ai/finetuner/pull/436))
+- Fix flaky integration test. ([#436](https://github.com/jina-ai/finetuner/pull/436))
 
 
 ## [0.1.0] - 2022-05-23